clean to single file
