+++ examples/acme_examples/.gitignore
@@ -1,0 +1,1 @@
+test*
