+++ examples/acme_examples/ppo_continuous.py
@@ -20,44 +20,75 @@ We are using the newest master version (344022e), so please make sure you
 install acme using method 4 (https://github.com/deepmind/acme#installation).
 """
 
+import logging
 import time
 from dataclasses import asdict
 from functools import partial
-from typing import Iterable, Iterator, List, Optional
+from typing import Any, Iterable, Iterator, List, Mapping, Optional, Union
 
-import acme_envpool_helpers.helpers as helpers
-import acme_envpool_helpers.lp_helpers as lp_helpers
+import dm_env
+import gym
 import jax
+import jax.numpy as jnp
 import numpy as np
 import reverb
 import tensorflow as tf
 import tree
 from absl import app, flags
-from acme import core, specs, types
+from acme import core, specs, types, wrappers
 from acme.adders import Adder
 from acme.adders.reverb import base as reverb_base
 from acme.adders.reverb import sequence as reverb_sequence
 from acme.agents.jax import actor_core as actor_core_lib
 from acme.agents.jax import actors, ppo
+from acme.agents.jax.actor_core import (
+  ActorCore,
+  FeedForwardPolicyWithExtra,
+  SimpleActorCoreStateWithExtras,
+)
 from acme.jax import experiments
 from acme.jax import networks as networks_lib
 from acme.jax import utils, variable_utils
+from acme.jax.types import PRNGKey
+from acme.utils.loggers import (
+  Logger,
+  LoggingData,
+  aggregators,
+  base,
+  filters,
+  terminal,
+)
+from dm_env import StepType
+
+import envpool
+from envpool.python.protocol import EnvPool
 
 FLAGS = flags.FLAGS
 
-flags.DEFINE_bool(
-  "run_distributed", False, "Should an agent be executed in a "
-  "distributed way (the default is a single-threaded agent)"
-)
-flags.DEFINE_integer("num_actors", 1, "Number of actors.")
 flags.DEFINE_string("env_name", "HalfCheetah-v3", "What environment to run.")
 flags.DEFINE_integer("seed", 0, "Random seed.")
 flags.DEFINE_integer("num_steps", 1_000_000, "Number of env steps to run.")
 flags.DEFINE_boolean("use_envpool", False, "Whether to use EnvPool.")
 flags.DEFINE_integer("num_envs", 8, "Number of environment.")
 flags.DEFINE_boolean("use_wb", False, "Whether to use WandB.")
+flags.DEFINE_string("wb_project", "acme", "WandB project name.")
 flags.DEFINE_string("wb_entity", None, "WandB entity name.")
 flags.DEFINE_string("desc", "", "More description for the run.")
+
+
+class TimeStep(dm_env.TimeStep):
+
+  extras: Any
+
+  def __new__(cls, step_type, reward, discount, observation, extras):
+    self = super(TimeStep,
+                 cls).__new__(cls, step_type, reward, discount, observation)
+    self.extras = extras
+    return self
+
+  def last(self) -> bool:
+    """Adapt the batch result for EnvironmentLoop episode-end checking."""
+    return any(self.step_type == StepType.LAST)
 
 
 class BatchSequenceAdder(reverb_sequence.SequenceAdder):
@@ -93,6 +124,73 @@ class BatchSequenceAdder(reverb_sequence.SequenceAdder):
     )
 
     return spec_step
+
+
+class AdderWrapper(Adder):
+
+  def __init__(self, adder: Adder) -> None:
+    self._adder: Adder = adder
+
+  def reset(self):
+    self._adder.reset()
+
+  def add_first(self, timestep: TimeStep):
+    if not any(timestep.first()):
+      raise ValueError(
+        'adder.add_first with an initial timestep (i.e. one for '
+        'which timestep.first() is True'
+      )
+    # Record the next observation but leave the history buffer row open by
+    # passing `partial_step=True`.
+    self._adder._writer.append(
+      dict(
+        observation=timestep.observation, start_of_episode=timestep.first()
+      ),
+      partial_step=True
+    )
+    self._adder._add_first_called = True
+
+  def add(
+    self,
+    action: types.NestedArray,
+    next_timestep: TimeStep,
+    extras: types.NestedArray = ...
+  ):
+    next_timestep = TimeStep(
+      step_type=next_timestep.step_type,
+      observation=next_timestep.observation,
+      reward=next_timestep.extras["reward"],
+      discount=next_timestep.discount,
+      extras=None,
+    )
+    self._adder.add(action, next_timestep, extras)
+
+
+def _batched_feed_forward_with_extras_to_actor_core(
+  policy: FeedForwardPolicyWithExtra
+) -> ActorCore[SimpleActorCoreStateWithExtras, Mapping[str, jnp.ndarray]]:
+  """Modified adapter allowing batched data processing."""
+
+  def select_action(
+    params: networks_lib.Params, observation: networks_lib.Observation,
+    state: SimpleActorCoreStateWithExtras
+  ):
+    rng = state.rng
+    rng1, rng2 = jax.random.split(rng)
+    action, extras = policy(params, rng1, observation)
+    return action, SimpleActorCoreStateWithExtras(rng2, extras)
+
+  def init(rng: PRNGKey) -> SimpleActorCoreStateWithExtras:
+    return SimpleActorCoreStateWithExtras(rng, {})
+
+  def get_extras(
+    state: SimpleActorCoreStateWithExtras
+  ) -> Mapping[str, jnp.ndarray]:
+    return state.extras
+
+  return ActorCore(
+    init=init, select_action=select_action, get_extras=get_extras
+  )
 
 
 class BuilderWrapper(ppo.PPOBuilder):
@@ -174,7 +272,7 @@ class BuilderWrapper(ppo.PPOBuilder):
     adder = super(BuilderWrapper, self).make_adder(replay_client)
     if not self._use_envpool:
       return adder
-    return helpers.AdderWrapper(adder)
+    return AdderWrapper(adder)
 
   def make_actor(
     self,
@@ -194,12 +292,74 @@ class BuilderWrapper(ppo.PPOBuilder):
       device="cpu",
       update_period=self._config.variable_update_period
     )
-    actor = helpers.batched_feed_forward_with_extras_to_actor_core(
-      policy_network
-    )
+    actor = _batched_feed_forward_with_extras_to_actor_core(policy_network)
     return actors.GenericActor(
       actor, random_key, variable_client, adder, backend="cpu"
     )
+
+
+class EnvPoolWrapper(wrappers.EnvironmentWrapper):
+
+  def __init__(self, environment: EnvPool):
+    super().__init__(environment)
+    self._num_envs = len(environment.all_env_ids)
+
+  def observation_spec(self) -> dm_env.specs.BoundedArray:
+    obs = self._environment.observation_spec().obs
+    new_obs = dm_env.specs.BoundedArray(
+      name=obs.name,
+      shape=[s for s in obs.shape if s != -1],
+      dtype="float32",
+      minimum=obs.minimum,
+      maximum=obs.maximum,
+    )
+    return new_obs
+
+  def reset(self) -> TimeStep:
+    ts = super().reset()
+    return TimeStep(
+      step_type=ts.step_type,
+      observation=ts.observation.obs,
+      reward=ts.reward,
+      discount=ts.discount,
+      extras=None,
+    )
+
+  def step(self, action) -> TimeStep:
+    ts = super().step(action)
+    ts = TimeStep(
+      step_type=ts.step_type,
+      observation=ts.observation.obs,
+      reward=ts.reward[0],  # Single value for recording the return.
+      discount=ts.discount,
+      extras={
+        # "step_type": ts.step_type,
+        "reward": ts.reward,
+      },
+    )
+    return ts
+
+
+def make_environment(task: str, use_envpool: bool = False, num_envs: int = 2):
+  env_wrappers = []
+  if use_envpool:
+    env = envpool.make(
+      task,
+      env_type="dm",
+      num_envs=num_envs,
+    )
+  else:
+    env = gym.make(task)
+    # Make sure the environment obeys the dm_env.Environment interface.
+    env_wrappers.append(wrappers.GymWrapper)
+  # Clip the action returned by the agent to the environment spec.
+  env_wrappers += [
+    partial(wrappers.CanonicalSpecWrapper, clip=True),
+    wrappers.SinglePrecisionWrapper
+  ]
+  if use_envpool:
+    env_wrappers.append(EnvPoolWrapper)
+  return wrappers.wrap_all(env, env_wrappers)
 
 
 def build_experiment_config():
@@ -217,7 +377,7 @@ def build_experiment_config():
 
   return experiments.Config(
     builder=ppo_builder,
-    environment_factory=lambda _: helpers.make_environment(
+    environment_factory=lambda _: make_environment(
       task,
       use_envpool=use_envpool,
       num_envs=num_envs,
@@ -230,31 +390,71 @@ def build_experiment_config():
   ), config
 
 
+def make_logger(
+  label: str,
+  steps_key: str = "steps",
+  task_instance: int = 0,
+  run_name: str = "",
+  wb_entity: str = "",
+  config: dict = {},
+) -> Logger:
+  del task_instance, steps_key
+  num_envs = config["num_envs"] if config["use_envpool"] else 1
+
+  print_fn = logging.info
+  terminal_logger = terminal.TerminalLogger(label=label, print_fn=print_fn)
+  loggers = [terminal_logger]
+  if label == "train":
+    import wandb
+    wandb.init(
+      project=FLAGS.wb_project,
+      entity=wb_entity,
+      name=run_name,
+      config=config,
+    )
+
+    class WBLogger(Logger):
+
+      def __init__(self, num_envs) -> None:
+        self._num_envs = num_envs
+
+      def write(self, data: LoggingData) -> None:
+        new_data = {}
+        for key, value in data.items():
+          if key in ["train_steps", "actor_steps"]:
+            key = "global_step"
+            value *= self._num_envs
+          new_data[key] = value
+        wandb.log(new_data)
+
+      def close(self) -> None:
+        wandb.finish()
+
+    loggers.append(WBLogger(num_envs))
+
+  # Dispatch to all writers and filter Nones and by time.
+  logger = aggregators.Dispatcher(loggers, base.to_numpy)
+  logger = filters.NoneFilter(logger)
+  logger = filters.TimeFilter(logger, 0.1)
+
+  return logger
+
+
 def main(_):
   experiment, config = build_experiment_config()
   if FLAGS.use_wb:
-    run_name = f"ppo__{FLAGS.env_name}"
+    run_name = f"acme_ppo__{FLAGS.env_name}"
     if FLAGS.use_envpool:
       run_name += f"__envpool-{FLAGS.num_envs}"
-    if FLAGS.run_distributed:
-      run_name += f"__dist-{FLAGS.num_actors}"
     if FLAGS.desc:
       run_name += f"__{FLAGS.desc}"
     run_name += f"__{FLAGS.seed}__{int(time.time())}"
     cfg = asdict(config)
     cfg.update(FLAGS.flag_values_dict().items())
     experiment.logger_factory = partial(
-      helpers.make_logger,
-      run_name=run_name,
-      wb_entity=FLAGS.wb_entity,
-      config=cfg
+      make_logger, run_name=run_name, wb_entity=FLAGS.wb_entity, config=cfg
     )
 
-  if FLAGS.run_distributed:
-    lp_helpers.run_distributed_experiment(
-      experiment=experiment, num_actors=FLAGS.num_actors
-    )
-  else:
     experiments.run_experiment(
       experiment=experiment,
       eval_every=experiment.max_number_of_steps,
