--- examples/acme_examples/setup.py
@@ -1,13 +1,0 @@
-import setuptools
-
-setuptools.setup(
-  name="acme-envpool-helpers",
-  version="0.1",
-  description="Examples of Acme using EnvPool.",
-  url="#",
-  author="Zichen Liu",
-  install_requires=[],
-  author_email="liuzc@sea.com",
-  packages=setuptools.find_packages(),
-  zip_safe=False
-)
